Remove separate extension package, get pdf rendering.
