--- packages/pdfviewer-extension/package.json
@@ -1,42 +1,0 @@
-{
-  "name": "@jupyterlab/pdfviewer-extension",
-  "version": "0.1.0",
-  "description": "JupyterLab - PDF Widget Extension",
-  "main": "lib/index.js",
-  "types": "lib/index.d.ts",
-  "files": [
-    "lib/*.d.ts",
-    "lib/*.js"
-  ],
-  "directories": {
-    "lib": "lib/"
-  },
-  "dependencies": {
-    "@jupyterlab/application": "^0.7.0",
-    "@jupyterlab/apputils": "^0.7.0",
-    "@jupyterlab/docregistry": "^0.7.0",
-    "@jupyterlab/pdfviewer": "^0.1.0"
-  },
-  "devDependencies": {
-    "rimraf": "^2.5.2",
-    "typescript": "~2.3.1"
-  },
-  "scripts": {
-    "build": "tsc",
-    "clean": "rimraf lib",
-    "watch": "tsc -w"
-  },
-  "jupyterlab": {
-    "extension": true
-  },
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/jupyterlab/jupyterlab.git"
-  },
-  "author": "Project Jupyter",
-  "license": "BSD-3-Clause",
-  "bugs": {
-    "url": "https://github.com/jupyterlab/jupyter-renderers/issues"
-  },
-  "homepage": "https://github.com/jupyterlab/jupyter-renderers"
-}
