+++ packages/pdfviewer/src/index.ts
@@ -2,21 +2,26 @@
 // Distributed under the terms of the Modified BSD License.
 
 import {
-  IInstanceTracker
-} from '@jupyterlab/apputils';
-
-import {
   Token
 } from '@phosphor/coreutils';
 
 import {
-  PDFViewer
+  ILayoutRestorer, JupyterLab, JupyterLabPlugin
+} from '@jupyterlab/application';
+
+import {
+  IInstanceTracker, InstanceTracker
+} from '@jupyterlab/apputils';
+
+import {
+  IDocumentRegistry
+} from '@jupyterlab/docregistry';
+
+import {
+  PDFViewer, PDFViewerFactory
 } from './widget';
 
 import '../style/index.css';
-
-export * from './widget';
-
 
 /**
  * A class that tracks PDF widgets.
@@ -32,3 +37,63 @@ interface IPDFTracker extends IInstanceTracker<PDFViewer> {}
 export
 const IPDFTracker = new Token<IPDFTracker>('jupyter.extensios.pdf-tracker');
 /* tslint:enable */
+
+/**
+ * The list of file extensions for PDFs.
+ */
+const EXTENSIONS = ['.pdf'];
+
+/**
+ * The name of the factory that creates PDF widgets.
+ */
+const FACTORY = 'PDF';
+
+/**
+ * The PDF file handler extension.
+ */
+const plugin: JupyterLabPlugin<IPDFTracker> = {
+  activate,
+  id: 'jupyter.extensions.pdf-handler',
+  provides: IPDFTracker,
+  requires: [IDocumentRegistry, ILayoutRestorer],
+  autoStart: true
+};
+
+
+/**
+ * Export the plugin as default.
+ */
+export default plugin;
+
+
+/**
+ * Activate the PDF widget extension.
+ */
+function activate(app: JupyterLab, registry: IDocumentRegistry, restorer: ILayoutRestorer): IPDFTracker {
+  const namespace = 'PDF-widget';
+  const factory = new PDFViewerFactory({
+    name: FACTORY,
+    modelName: 'base64',
+    fileExtensions: EXTENSIONS,
+    defaultFor: EXTENSIONS,
+    readOnly: true
+  });
+  const tracker = new InstanceTracker<PDFViewer>({ namespace });
+
+  // Handle state restoration.
+  restorer.restore(tracker, {
+    command: 'file-operations:open',
+    args: widget => ({ path: widget.context.path, factory: FACTORY }),
+    name: widget => widget.context.path
+  });
+
+  registry.addWidgetFactory(factory);
+
+  factory.widgetCreated.connect((sender, widget) => {
+    // Notify the instance tracker if restore data needs to update.
+    widget.context.pathChanged.connect(() => { tracker.save(widget); });
+    tracker.add(widget);
+  });
+
+  return tracker;
+}
