+++ packages/pdfviewer/style/index.css
@@ -4,4 +4,6 @@
 |----------------------------------------------------------------------------*/
 
 .jp-PDFViewer {
+  width: 100%;
+  height: 100%;
 }
